requirements.txt
moviepy
pydub
gtts
openai-whisper
ffmpeg